snap_qp: Fix coverity issue of potential address overflow
cqe_size and cqe_cnt are both 32 bit size, and multiplied before
assigned as a 64-bit dbr_addr.
One of them must be cast to 64-bit number,to avoid overflow.

+++ src/snap_qp.c
@@ -92,7 +92,7 @@ static int devx_cq_init(struct snap_cq *cq, struct ibv_context *ctx, const struc
 		umem_id = devx_cq->devx.umem.devx_umem->umem_id;
 		umem_offset = 0;
 		dbr_umem_id = umem_id;
-		dbr_addr = attr->cqe_size * devx_cq->cqe_cnt;
+		dbr_addr = (uint64_t)attr->cqe_size * devx_cq->cqe_cnt;
 	} else {
 		if (attr->dpa_element_type == MLX5_APU_ELEMENT_TYPE_THREAD) {
 			if (!attr->dpa_thread) {
